Reworked the properties bar. Instead of sliders and textfields, a customized version of Numberbox is now used. Removed most of the dirty hacks (e.g. dummySliders).
The value of a NumberInputBox can be changed both by clicking on the
NumberInputBox and dragging the mouse to the left (lower values) or the
right (higher values) and by clicking on the NumberInputBox and using
the keyboard to enter a value. The only accepted input are numbers as
well as backspace and return.

Furthermore, an interface Pluggable has been added. Connection, Cutout,
STLMesh and Shape now implement this interface. The properties bar can
now be plugged into any object that implements Pluggable. Thus, we don't
need to have a gazillion methods for each different Type.